Fix switchdev transition after configuring 256 SFs
The previous backporting of upstream commit 1fb7f8973f51 ("RDMA: Support
more than 255 rdma ports") missed one hunk. Fix this.

Orabug: 33913142

Fixes: 4aacc5eec85d ("RDMA: Support more than 255 rdma ports")

+++ drivers/infiniband/core/cache.c
@@ -1457,7 +1457,7 @@ err:
 }
 
 static int
-ib_cache_update(struct ib_device *device, u8 port, bool update_gids,
+ib_cache_update(struct ib_device *device, u32 port, bool update_gids,
 		bool update_pkeys, bool enforce_security)
 {
 	struct ib_port_attr       *tprops = NULL;
